feat: `sim.State`
A simple key/value in-memory state for the simulator.

Use the preflight method `token(key)` to obtain a token that can be used to reference the value of the state at runtime.

During deployment (i.e. `cloud.OnDeploy` or `cloud.Service` startup), you must call the inflight method `set(key, value)` to set the runtime value. The value will be available at runtime through the inflight method `get(key)` (or resolved as a token).

See tests for examples.