feat: `sim.State` (#4561)
A simple key/value in-memory state for the simulator with support for lazy tokens.

Consider a use case where there is an attribute of a simulated service that only gets resolved during initialization (e.g. the exposed port of a container). In order to create such resources, we need a way to obtain a lazy token that gets resolved during simulator initialization. It just so happens that we already have a mechanism like this, but it was not exposed as a public API.

Use the preflight method `token(key)` to obtain a token that can be used to reference the value of the state at runtime.

During deployment (i.e. `cloud.OnDeploy` or `cloud.Service` startup), you must call the inflight method `set(key, value)` to set the runtime value. The value will be available at runtime through the inflight method `get(key)` (or resolved as a token).

See tests for examples.

Inspired by #3793

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.