chore(sdk): force usage of mime v3 throughout repo (#4590)
We were already depending on `mime`, so this does not increase our dependency closure.

The version of `mime` that `express` depends on (via the `send` package) is very old and can easily conflict with other dependencies we may want to add. This forces our packages to always use v3 of `mime`, and patches `mime` v3 to expose the previously removed functions.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.