fix(compiler): can't inherit an stdlib resource (#4640)
Fixes #3585

Before this PR our dependency injection only worked on `new` calls. So when instantiating an stdlib `cloud.Api` for example you'd get a `cloud.Api` instance matching the desired target. But when inheriting from an stdlib preflight class dependency injection wasn't being done. So we just inherited the abstract base class. 
With this change we get dependency injection also when doing `extends cloud.SomeThing`. The way I got this to work is by changing the dep-injection mechanism to return not a concrete instance but rather a concrete type. Then we can instantiate that type or inherit it based on what we need.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.