fix(vscode): many completion improvements (#4607)
Was traveling recently and did random in-transit work to improve completions. Sorry for so many fixes at once, hopefully the improvement is worth it.

- Fixes #4295
signature_help does a better job of getting the inner-most function call

- Fixes #4000
Completions now filters out function that return void unless you're at the scope-level statement (e.g. it won't show up when you're expecting a plain old expression)

- Fixes #3440
The `inflight` transformer was the issue here due to its injected scope. The visitor handles this better now.

- Fixes #3993
This was really the same bug as #3440

- Fixes #3994
Partially the same bug as #3440, but also the completion thought you wanted a type because it was following a ":". Both are fixed.

- Fixes #1984
Simply made sure not to include any symbols that were defined after the current scope. Used spans for this instead of stmt indexes.

- Fixes #3990
Adds `bring` when looking for completions in the top scope

- Fixes #4013
Simplified logic to lookup types, kinda ended up fixing this by accident

- Fixes #4459
Basically the same bug as #4013

- Fixes #3995
Technically the inflight snippet is a valid completion item to provide to a struct, so I didn't remove it, but made sure to deprioritize snippets like that

- Fixes #4001
Added completions for builtin types

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.