fix(compiler): jsii import use-before-defined error (#4657)
Fixes #4575

This was a tricky bug to pin down. I initially thought it might be an issue with our symbol lookup code, but it turns out it was working fine - it instead there was an issue in our symbol environment construction. In `class2.w`, the `util` symbol had the was being stored with the span and statement index information associated with `class1.w`.

The root cause of this seems to be that after the first `bring util` statement, some of the import information is stored in `JsiiImportSpec`, which then got reused when `bring util` was called a second time. To fix this, I updated the logic we perform for bring-jsii statements so that it only reuses an existing `JsiiImportSpec` if the `alias` it has stored is exactly the same (both the string and its position in the source code).

I'm not sure if storing a separate `JsiiImportSpec` for every unique bring statement has any performance drawbacks but I didn't have enough context to easily refactor the `alias` and `import_statement_idx` fields out of `JsiiImportSpec`:

https://github.com/winglang/wing/blob/392aa4bdce6868ad01ac5df674a7decb0a823f21/libs/wingc/src/type_check/jsii_importer.rs#L51-L69

Definitely interested if there are ways to simplify this though.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.