feat(sdk): support multiple test simultaneously on awscdk (#4654)
When creating a stack with the AWS CDK, it's necessary to set the environment variable CDK_STACK_NAME. This variable is causing a problem when trying to execute multiple tests simultaneously.

The solution I found was to capture this variable and concatenate it with the filename. This way, each stack will be named for deployment when running tests in the format:

CDK_STACK_NAME + "-" + file name

Closes #3750 

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.