fix(compiler): `async` method modifier never used (#4674)
The `async` method modifier wasn't being used anywhere in the code and wasn't marked as  reserved word in parser.rs anyway.
The wing way is basically that everything inflight is `async` anyway and everything preflight isn't.
The only thing I'm not sure of is whether we should add `async` as a reserved word just as a space-holder. I think not.

This is a minor cleanup related to #3329.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always) (fixed grammar test)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.