chore(sdk): allow aws sdk endpoint configuration via env variables (#4720)
This upgrades the aws sdk packages to allow configuration of [endpoints via env](https://aws.amazon.com/blogs/developer/new-improved-flexibility-when-configuring-endpoint-urls-with-the-aws-sdks-and-tools/), mostly for Localstack, but might be useful in other contexts as well. Looks like this got added very recently to the JS AWS SDK v3 (https://docs.aws.amazon.com/sdkref/latest/guide/feature-ss-endpoints.html)

In addition, this gets rid of the deprecated packages `@aws-sdk/util-stream-node` / `@aws-sdk/util-utf8-node` and replaces them with the `@smithy/...` equivalent. 

## Checklist

- [ ] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [ ] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.