fix(sdk): adding multiple files with awscdk addObject (#4738)
For each file we want to send to S3, it creates a CloudFormation resource of type `Custom::CDKBucketDeployment`.

The problem is that this resource expects an array of files to be placed in S3 and doesn't allow for more than one of these resources in the CloudFormation template.

I modified the code of addObject to create only one resource and include the files in this resource

Closes #4737 

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.