feat(compiler): bring trusted libraries (#4795)
Introduces the capability to bring trusted libraries using the `bring module_name;` syntax, as a short-hand for `bring "@winglibs/module_name" as module_name;`. If the module hasn't been installed, it will be a parsing error (as it's not possible to type check usage of that module until it has been installed).

Part of #1037

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.