fix(cli): wing pack doesn't pack subdirectories (#4798)
Fixes #4797

The root cause of this bug was that `wing pack` was incorrectly globbing files to copy into its staging directory. Instead of reinventing the wheel, I've changed the implementation to simply reuse the npm tooling that determine what files are included in a npm tarball, and use that to determine what should get copied into the staging directory.

I've also updated `wing pack` so it runs `wing compile` inside the staging directory instead of inside the user's original directory to reduce the chance for compile errors seeping into packaged Wing libraries.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.