feat(vscode): goto file in bring statement (#4910)
Closes #4899

Logic is the same as the extern modifier.
The behavior for a directory is a bit awkward but I decided to leave it in. Ideally, it'd be nice if it just "revealed" the dir in the file explorer, but that'd have to be implemented on a per-editor basis and not part of the LSP.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.