feat: `Node.of(this).app` and `.root`
Makes it easy to access the root app from every node through `Node.of(this).app`.

`Node.of(this).root` is equivalent. It also returns the same root app, even if it's not the absolute root of the tree (e.g. in case the app is part of a surrounding testing tree).

This also makes it easier to implement app-level singletons:

```js
let uid = "unique-id";
let root = Node.of(this).root;
return unsafeCast(root.tryFindChild(uid)) ?? new MySingleton() as uid in root;
```

P.S. In the future, this is supposed to become `nodeof(this).app`.

Fixes #513

TODO:
- [ ] Add a doc topic about singletons.