feat: singletons + enable creation of preflight objects in static (#5029)
## Support the singleton pattern

Resolves #513 by making it easy to access the root app from every node through `Node.of(this).app` or `Node.of(this).root`. Both return the same root app, even if it's not the absolute root of the tree in case the app is part of a surrounding testing tree.

> In the future, this is supposed to become `nodeof(this).app`.

Also add `tryFindChild(id)` to the app to make it easier to implement app-level singletons:

```js
let uid = "unique-id";
let root = Node.of(this).root;
return unsafeCast(root.tryFindChild(uid)) ?? new MySingleton() as uid in root;
```

Added a doc example with this information.

Fixes #5015

## Enable creation of preflight classes in static methods

Fix #2583 by requiring `in scope` when creating preflight objects in static methods:

```js
class Foo {
  static maker(scope: std.IResource): Foo {
    return new Foo() in scope;
  }
}
```

Also, emit an error if there is no explicit scope.

```js
class Foo {
  static maker(): Foo {
    return new Foo();
//         ^^^^^^^ Cannot instantiate without an explicit scope
  }
}
```

This is also working for JSII classes and resolves #4987. It required a modification on how we emit `this.node.new()` to use the explicit scope instead of `this` in this case.

Resolves #4246 by allowing any `IConstruct` to be used within `in scope`.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.