fix: singletons are not visible in the console
Change `Node.of().root` to return the root of the user app instead of the app. This is needed because the user app is actually placed inside a direct child of the app and not the app itself.

Since the Wing Console only shows the contents of this direct child, then we want to mount singletons on that "root" instead of the app itself.