fix(sdk): external platforms may have error when lifting resources (#5069)
Fixes #5064

`instanceof` strikes again.

Technically hangar already tests the case that failed here, but the test did not fail in hangar. I don't actually have an answer for why this is. My guess is that the cwd hangar runs in affected the modules loaded.

I did not add any testing here because afaik we are not currently equipped to test this scenario properly (we need https://github.com/winglang/wing/issues/3390), so I only tested that this fix worked locally.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.