fix!: singletons are not visible in the console (#5047)
Change `Node.of().root` to return the root of the user app instead of the app. This is needed because the user app is actually placed inside a direct child of the app and not the app itself.

Since the Wing Console only shows the contents of this direct child, then we want to mount singletons on that "root" instead of the app itself.

BREAKING CHANGE: `node.root` now returns `IConstruct` instead of `IApp`.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.