fix(sdk): awaiting client object causes error in sim target (#5120)
I was debugging an error with @yoav-steinberg, and we had an issue while augmenting the lifting system to get the following code to work:

```js
let b = [new cloud.Bucket()];
test "foo" {
  b.at(0);
}
```

We found that the generated JavaScript looked like this:

```js
await b.at(0);
```

Ordinarily this would no-op (since calling "await" on non-promises usually does nothing). But in the Wing simulator, the bucket client is a special proxy object that automatically translates all method calls into HTTP requests. When determining if a value is a promise or not, the JS runtime tries to call "then" on objects (see https://stackoverflow.com/questions/55262996/does-awaiting-a-non-promise-have-any-detectable-effect), and this ended up causing an error instead of the no-op behavior.

The first major bug was that trying to call a non-existent method on the proxy object was crashing the simulator server, so that's fixed.

Beyond that, to accommodate the "no-op await" use case, I've added some special case handling so if the the client contacts the server, trying to call `then()`, and the server responds saying the method does not exist, then the client object will return itself back. Of course, we probably shouldn't generate `await b.at(0)` since `at` is a synchronous method on `Array`, but I don't think this does any harm.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.