feat(sdk): simulator state (#5164)
This PR adds the capability for the Wing simulator to store resource state across compilations. This means if you performed a lot of work to fill a `cloud.Bucket` with objects or values in `cloud.Counter`, it will be remembered when you recompile and reload the application as long as the resource is still in your app. This will be extended to other resources like Queue and Workload in the future.

Here's a screen recording - notice how even as the app is updated with new resources or different inflight code, the state of resources doesn't go away.

https://github.com/winglang/wing/assets/5008987/a877e0f1-8085-4264-bf32-1ac7c971af4c

Notes:
- The cloud.TestRunner resource is now only added to a Wing app's construct tree when it is compiled in test mode (with `wing test`).
- The Wing Console now runs two copies of the app, one compiled in the default mode, and one compiled with tests. The resource graph only shows the default app, but the tests panel shows a list of all tests from the app compiled with the simulator, and clicking any of the tests will run them with fresh state every time.

Future work:

- https://github.com/winglang/wing/issues/5197
- https://github.com/winglang/wing/issues/5198

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.