fix(sdk): no resource found with handle simulator error (#5229)
Fixes #5228

We discovered an issue where invalid errors were thrown in unit tests where a message is pushed to a queue and the queue had a consumer function. The problem occurred because a queue's simulation periodically processes messages, but this continued even after the queue consumer function is deleted during the simulator's cleanup. (I've shared some verbose test logs from the minimal repro in #5228 here: https://gist.github.com/Chriscbr/23fb248fb65d9e349da6e3347c90c866).

To fix this, I updated the EventMapping resource so it also removes the its function from the queue's subscribers when the EventMapping is cleaned up. By doing this, we guarantee the function is removed from the queue's subscribers before the function is cleaned up, so when the queue tries processing messages, it will send the messages to another consumer or simply retain them in memory.

Misc:
- I also cleaned up some dead code that was previously used by the CLI for picking out distinct tests to run (no longer needed after https://github.com/winglang/wing/pull/5164)

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.