fix(compiler)!: doesn't emit error for missing return statements (#5249)
Fixes #4187
Closes #4972
Fixes #1843

This fixes a gap in Wing where no error would be thrown if you forgot to `return` in a function whose return type expects one. Now, the compiler requires a `return` statements in most functions. The exceptions are:
- Functions that return `void` may omit the `return;` statement
- Functions that return `T?` (an option value) may omit the `return <value>;` or `return nil;` statement - the function will implicitly return `nil`

The check is not comprehensive, as the compiler isn't able to validate all code paths will return (for example the `return` statement could be inside of an if branch, in which case your function would still have a bug), but it should catch many basic errors.

The PR also resolves an issue where calling `invoke()` on `cloud.Function` in the simulator could leak a JavaScript "undefined" value. As a solution, the simulator falls back to returning an empty string in these cases. (This was already the existing behavior for the `tf-aws`, `tf-gcp`, and `tf-azure` platforms).

BREAKING CHANGE: Functions with a non-void and non-option return type must return a value. If you are passing an inflight function to `new cloud.Function()` and you don't want to return a value, you can fix your code by changing the return type to `str?`, or omitting the return type annotation so it's inferred:

```js
bring cloud;

new cloud.Function(inflight (): str? => {
  log("this works");
});

new cloud.Function(inflight () => {
  log("this also works");
});
```

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.