feat: use typescript with wing cli and worker thread for preflight (#5247)
#### General Vibe
- This is an MVP of getting TS working via the wing cli (`compile`, `run`, and `test`). This is basically only usable for preflight
- I wanted to change as little as possible in the SDK to avoid wasted work once resources are moved to winglibs
- Does not address inflight stuff yet

The entrypoint TS experience is different than the existing issue/story:

```ts
import { main } from "ts4wing";

main((app) => {
  // stuff goes here
})
```

This was due to our usage of a "root" construct for testing isolation. The wing cli needs to have control over the instantiation of resources, so `new App()` is not really feasible because we don't want users to actually attach stuff at the app-level. It may be possible to get it working with `new App()` but the hacks needed didn't seem worth it.
An added benefit of this approach is that the user no longer needs to call `app.synth()`

Anecdotally, I have built an awscdk-based internal framework at a previous job that looked like this for a similar reason (to duplicate stuff across Stages) and it worked pretty well.

#### Indirect Changes
- Made sure everything matches their `@types/node` version
- **preflight execution now happens in a worker thread instead of a VM**
  - VM context did not allow data to be provided past the initially loaded file.
  - In a VM, process.env is useful but pollutes the parent process
  - The overloaded "require" was only useful for the initial file. Now I added a shim that overloads require everywhere to make sure the SDK is loaded properly

Fixes #3678

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.