feat: emit extern paths as relative (#5361)
This change takes advantage of the assumption that the output directory will always be a subdirectory of the entrypoint (or entrydir for lib). This also handles paths in wing libraries, which will be emitted as a barrel import instead.

This is mostly important for https://github.com/winglang/wing/pull/5278, but it's also generally nice to avoid absolute paths when they aren't needed

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.