fix(compiler): can't pass inflight arg to phase independent method on lifted object (#5158)
see #3593
When the type checker needs to assign a concrete phase to a phase independent property instead of using the object's phase we now take into account cases where we need the property to be inflight even if the object is a lifted preflight instance. If the phase independent method receives any inflight inputs it'll need to be executed inflight and therefore treated as an inflight method. This means that if the object is an inflight object **or** if the property is a method call and any of the arguments passed are inflight expressions then the property will be treated as an inflight property.

## Additional changes
### Mechanism for lifting non resource std lib classes to inflight
Once I did the above change we encountered problem when accessing some basic std types inflight (Duration/Datetime) because now only part of the type was lifted instead of the entire expressions:
```wing
let x = datetime.systemNow();
test "mytest" { 
  x.toIso() // Now only `x` is lifted and `toIso()` is executed inflight
}
```
The problem is that `Datetime` objects weren't being lifted. To fix this I added a mechanism where we can mark std classes as `ILiftable` even if they aren't a resource. This means that when used the std lib code is bundled into inflight and the relevant object (`Datetime`/`Duration`) is serialized into inflight and deserialized inflight and can be used freely.
This also cleans up special case lifting code out of `_liftObject` which had some special cases for `Duration` which were broken anyway (for example when trying to access `milliseconds` field in a `Duration`).

### Better error when accessing preflight members of inflight (lifted) object:
`Can't access preflight member \"{property}\" on inflight instance of type \"{instance_type}\"`
Previously `lifted_object.preflight_field` would pass type checking and error duing the lifted code with a more general (can't mix inflight expression in preflight expression) error.

### Calling phase independent static methods with inflight args
This is now also supported and will result in the phase independent method to be executed inflight.

### Binary expression / string interpolation evaluation returns correct expr phase
Before something like `a + b` would either return the phase of `a` or `b` or `Phase::Independent` (depending on the type of `a` and `b` and which operator was used). This was pretty random. We now return the appropriate phase based on the phases of `a` and `b` using a new `combine_phases` function. This was required so when such expressions are passed as arguments to a method we need to know if the resulting expression can be lifted or includes inflight components in it.

### More correct errors on failure to qualify a lift
We now report a diagnostic when we encounter access to a property of an **inflight** expression that evaluates to an object of a preflight class type. This is more correct than reporting a qualification error on any non-property reference to a preflight object, which wasn't good enough anymore given the other changes in this PR. Details in this comment: https://github.com/winglang/wing/pull/5158#discussion_r1435802072

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.