fix: `_toInflight()` fails in typescript when consuming a wing library (#5460)
**Note**: This PR is a setup for the upcoming typescript PR. This change required many snapshot updates so it made sense to keep it contained in a separate PR.

Normally in wing, all preflight code is executed in the same folder. This typically also works fine for typescript, except when calling `_toInflight()` on a inflight closure class because it returns a string with a relative require. This works fine in winglang but not typescript because we execute that string elsewhere. Adding __dirname ensures that the require will resolve correctly.

Added a helper to make sure windows paths don't break everything and normalized helper usage a little bit.

This usecase currently doesn't exist in wing, thus the lack of a test. A followup PR will make use of this.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.